fix: fail on invalid policy lines instead of silently skipping
- Move validate_policy() from cli.py to parser.py for reuse
- Add policy validation at proxy startup
- In enforcement mode: fail if policy has invalid lines
- In audit mode: emit warning annotation and continue

This prevents typos in policy rules from silently removing allow rules,
which could cause unexpected blocks or false security assumptions.

Fixes #50
